forge_output/warehouseflow/workspace/requirements.txt
alembic==1.13.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
pytest==8.0.2
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.9
requests==2.31.0
pytest-benchmark==4.0.0
locust==2.32.3
